requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
docopt==0.6.2
dominate==2.6.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
pipreqs==0.4.11
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.27.1
SQLAlchemy==1.4.32
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.1
yarg==0.1.9
